♻️ Refactor: guard 위치 수정 (#96)
