+++ src/main/nest/domain/auth/auth.module.ts
@@ -7,7 +7,7 @@ import { LocalStrategy } from './strategy/local.strategy';
 import { JwtStrategy } from './strategy/jwt.strategy';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { NaverStrategy } from './strategy/naver.strategy';
-import { NaverAuthGuard } from './guard/naver-auth.guard';
+import { NaverAuthGuard } from '../../global/guard/naver-auth.guard';
 
 @Module({
   imports: [
