+++ src/main/nest/global/decorator/api-guard-optional.decorator.ts
@@ -1,6 +1,6 @@
 import { applyDecorators, UseGuards } from '@nestjs/common';
 import { ApiBearerAuth } from '@nestjs/swagger';
-import { JwtAuthOptionGuard } from '../../domain/auth/guard/jwt-auth-optional.guard';
+import { JwtAuthOptionGuard } from '../guard/jwt-auth-optional.guard';
 
 export function ApiOptionGuard() {
   return applyDecorators(UseGuards(JwtAuthOptionGuard), ApiBearerAuth('accessToken'));
